requirements.txt
torch>=1.9.0
torchvision>=0.10.0
segmentation-models-pytorch>=0.3.0
opencv-python>=4.5.0
albumentations>=1.0.0
matplotlib>=3.3.0
pillow>=8.0.0
numpy>=1.20.0
tqdm>=4.60.0
timm>=0.4.0
